Badges and Daily Logs and Stats
The verification method was added to the dashboard_fragment class. The method checks if the user meets any requirements to gain any badges and uses a toast to notify them if they do.

The day to day loging functionality has been added by saving the last date the app was opened and comparing it to the current date. If its diffrent the new date takes its place and the UI and logs are updated to reflect so.

The stats page has also been created and inflated in its fragment.

ToDo: The badge fragment needs to be functional and needs to display badges to reflect the users collection. The stats page should display the stats collected by the app and the email sending function needs to be added.