Working Email Sending of logs and code refractoring
The logs can now be sent when a button is clicked on the dayview of the dashbaord. This button will launch an intent and populate an email with the logs in userfriednly readble text.

The code has also been refractored to be more organised and eaiser to maintain.

Todo: The reminders need fixing as currently no backend exists for them. The app needs to account for the state changes as currently the rotation of the device will cause a crash. Following these changes the prototype should be completed.